refactor(persona): Componentize and comment persona management
This commit refactors the persona management feature to improve code organization, maintainability, and comprehension.

Key changes:
- Created a new, dedicated `PersonasPage.jsx` component to encapsulate all logic and UI for persona management.
- Moved all persona-related state, handlers, and effects from the monolithic `HomePage.jsx` to the new `PersonasPage.jsx`.
- Simplified `HomePage.jsx` by removing the extracted logic and rendering the `PersonasPage` component when its view is active.
- Added comprehensive JSDoc comments to the new component, the PersonaWizard, the backend API files (`api/personas/`), and the `personaState.js` utility file to clarify their functionality.

This refactoring addresses technical debt from the previous migration from localStorage to a PostgreSQL backend, making the feature more modular and easier to maintain.
