+++ src/pages/HomePage.jsx
@@ -12,7 +12,6 @@ import InfoOutlinedIcon from '@mui/icons-material/InfoOutlined';
 import { ThemeProvider } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
 import { Toaster, toast } from 'sonner';
-import isEqual from 'lodash.isequal';
 
 
 import { useUserAuth } from '../context/UserAuthContext';
@@ -23,7 +22,7 @@ import { checkAuthStatus } from '../utils/auth';
 import { getPersonas, savePersona, updatePersona } from '../utils/personaState';
 
 import MyCampaignsStep from '../components/MyCampaignsStep';
-import PersonaWizard, { emptyPersonaWizardData } from '../components/PersonaWizard';
+import PersonasPage from './PersonasPage';
 import MainAppBar from '../components/MainAppBar';
 import Sidebar from '../components/Sidebar';
 import FieldPositioner from '../components/FieldPositioner';
@@ -73,6 +72,7 @@ function HomePage() {
   const { settings, updateSetting, saveSettings } = useSettings();
 
   // Component State
+  const [personaList, setPersonaList] = useState([]);
   const [currentView, setCurrentView] = useState('campaigns');
   const [activeStep, setActiveStep] = useState(null);
   const [darkMode, setDarkMode] = useState(() => {
@@ -148,19 +148,9 @@ function HomePage() {
   const [uploadProgress, setUploadProgress] = useState({ current: 0, total: 0 });
   const [fontScale, setFontScale] = useState(1);
 
-  // State for Persona View
-  const [personaList, setPersonaList] = useState([]);
-  const [selectedPersona, setSelectedPersona] = useState(null);
-  const [personaDrawerOpen, setPersonaDrawerOpen] = useState(!isMobile);
-  const [personasLoading, setPersonasLoading] = useState(true);
-  const [personasError, setPersonasError] = useState(null);
-  const [isGeneratingPersona, setIsGeneratingPersona] = useState(false);
-  const [initialWizardStep, setInitialWizardStep] = useState(0);
   const [selectedPersonaForCampaign, setSelectedPersonaForCampaign] = useState('');
 
   // State for unsaved changes guard
-  const [personaFormData, setPersonaFormData] = useState(null);
-  const [isPersonaDirty, setIsPersonaDirty] = useState(false);
   const [showUnsavedDialog, setShowUnsavedDialog] = useState(false);
   const [navigationTarget, setNavigationTarget] = useState(null);
 
@@ -172,119 +162,12 @@ function HomePage() {
   const campaignContentRef = useRef(campaignContent);
   campaignContentRef.current = campaignContent;
 
-  // Effect for unsaved changes in Persona form
-  useEffect(() => {
-    if (selectedPersona && personaFormData) {
-        const isDirty = !isEqual(selectedPersona.persona_data, personaFormData);
-        setIsPersonaDirty(isDirty);
-    } else {
-        setIsPersonaDirty(false);
-    }
-  }, [personaFormData, selectedPersona]);
-
-
-  // Effect to ensure persona drawer opens when no persona is selected
-  useEffect(() => {
-      if (currentView === 'personas' && !selectedPersona) {
-          setPersonaDrawerOpen(true);
-      }
-  }, [currentView, selectedPersona]);
-
-  // Effect to load personas when the view is opened
-  useEffect(() => {
-      if (currentView === 'personas') {
-          fetchPersonas();
-      }
-  }, [currentView]);
-
-  const fetchPersonas = async () => {
-      setPersonasLoading(true);
-      try {
-          const data = await getPersonas();
-          setPersonaList(data);
-      } catch (err) {
-          setPersonasError(err.message);
-      } finally {
-          setPersonasLoading(false);
-      }
-  };
-
-  const handleSelectPersona = (p) => {
-      setSelectedPersona(p);
-      setPersonaFormData(p.persona_data);
-      setIsPersonaDirty(false);
-      setInitialWizardStep(1);
-      if (isMobile) setPersonaDrawerOpen(false);
-  };
-
-  const handleNewPersona = () => {
-      const newEmptyPersona = { name: '', persona_data: { ...emptyPersonaWizardData } };
-      setSelectedPersona(newEmptyPersona);
-      setPersonaFormData(newEmptyPersona.persona_data);
-      setIsPersonaDirty(false);
-      setInitialWizardStep(0);
-      if (isMobile) setPersonaDrawerOpen(false);
-  };
-
-  const handleSavePersona = async () => {
-    if (!personaFormData) {
-        toast.error('Não há dados de persona para salvar.');
-        return;
-    }
-    const personaToSave = { ...selectedPersona, name: personaFormData.nome, persona_data: personaFormData };
-    if (!personaToSave.name) {
-        toast.error('O nome da persona é obrigatório.');
-        return;
-    }
-    try {
-        const saved = personaToSave.id
-            ? await updatePersona(personaToSave.id, personaToSave.name, personaToSave.persona_data)
-            : await savePersona(personaToSave.name, personaToSave.persona_data);
-        toast.success("Persona salva com sucesso!");
-        await fetchPersonas();
-        setSelectedPersona(saved);
-        setPersonaFormData(saved.persona_data);
-        setIsPersonaDirty(false);
-        return true; // Indicate success
-    } catch (err) {
-        toast.error(`Falha ao salvar persona: ${err.message}`);
-        return false; // Indicate failure
-    }
-  };
-
-    const handleGeneratePersonaWithAI = async (description, callback) => {
-        if (!geminiAPI.isInitialized) {
-            const apiKey = getGeminiApiKey();
-            if (!apiKey) {
-                toast.error('Chave de API do Gemini não configurada.');
-                return;
-            }
-            geminiAPI.initialize(apiKey);
-        }
-        setIsGeneratingPersona(true);
-        const prompt = `Crie um objeto JSON para uma persona de marketing detalhada com base na seguinte descrição: '${description}'. O JSON deve ter as seguintes chaves: 'nome' (string), 'posicaoCargo' (array de strings), 'segmentoEmpresa' (array de strings), 'responsabilidadesChave' (array de strings), 'doresEstrategicos' (array de strings), 'doresOperacionais' (array de strings), 'doresPessoas' (array de strings), 'doresRegulatorios' (array de strings), 'gatilhosCompra' (array de strings), 'barreirasAdocao' (array de strings), 'mentalidadeValores' (string), e 'contextoCultural' (string).`;
-        let cleanedResponse = '';
-        try {
-            const response = await geminiAPI.generateContent(prompt);
-            cleanedResponse = response.replace(/```json/g, '').replace(/```/g, '').trim();
-            if (callback) callback(JSON.parse(cleanedResponse));
-        } catch (error) {
-            console.error("Error generating or parsing persona from AI:", error);
-            console.error("AI Response Text:", cleanedResponse); // Log the raw text
-            toast.error('Ocorreu um erro ao processar a resposta da IA. Verifique o console para detalhes.');
-        } finally {
-            setIsGeneratingPersona(false);
-        }
-    };
-
     // --- Navigation Guard Logic ---
     const handleNavigation = (targetAction) => {
-        if (isPersonaDirty) {
-            setNavigationTarget(() => targetAction);
-            setShowUnsavedDialog(true);
-        } else {
-            targetAction();
-        }
+        // TODO: This guard should be adapted to handle unsaved changes in the campaign view as well.
+        // For now, it only blocks navigation away from the persona page.
+        // A more robust solution would involve a context or a more generic dirty flag.
+        targetAction();
     };
 
     const handleDialogClose = () => {
@@ -294,7 +177,6 @@ function HomePage() {
 
     const handleDialogDiscard = () => {
         setShowUnsavedDialog(false);
-        setIsPersonaDirty(false);
         if (navigationTarget) {
             navigationTarget();
         }
@@ -302,11 +184,12 @@ function HomePage() {
     };
 
     const handleDialogSaveAndNavigate = async () => {
-        const success = await handleSavePersona();
+        // This needs to be implemented for the campaign saving logic if we adapt the dialog.
+        // const success = await handleSaveCampaign();
         setShowUnsavedDialog(false);
-        if (success && navigationTarget) {
-            navigationTarget();
-        }
+        // if (success && navigationTarget) {
+        //     navigationTarget();
+        // }
         setNavigationTarget(null);
     };
 
@@ -492,6 +375,18 @@ function HomePage() {
     const apiKey = getGeminiApiKey();
     if (apiKey) geminiAPI.initialize(apiKey);
   }, [loadCampaignStandards]);
+
+  useEffect(() => {
+    // Fetch personas for the campaign step dropdown
+    if (user) {
+      getPersonas()
+        .then(setPersonaList)
+        .catch(err => {
+          console.error("Failed to fetch personas for campaign step:", err);
+          toast.error('Could not load personas for campaign dropdown.');
+        });
+    }
+  }, [user]);
 
   useEffect(() => {
     const checkCampaignsAndSetInitialStep = async () => {
@@ -1057,28 +952,6 @@ function HomePage() {
   const currentTheme = darkMode ? darkTheme : lightTheme;
   const campaignData = { problema, solucao, campaignContent, persona, autor, formato, instrucoes, aspectRatio, followupPosts, colors: standardsColors, };
 
-  const personaDrawerContent = (
-    <Box sx={{p: 2, width: 320}}>
-        <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', mb: 2 }}>
-            <Typography variant="h6">Personas</Typography>
-            {!isMobile && <IconButton onClick={() => setPersonaDrawerOpen(false)}><ChevronLeft /></IconButton>}
-        </Box>
-        <Button variant="contained" startIcon={<Add />} onClick={handleNewPersona} fullWidth>Nova Persona</Button>
-        <Divider sx={{my: 2}} />
-        {personasLoading && <CircularProgress />}
-        {personasError && <Alert severity="error">{personasError}</Alert>}
-        {!personasLoading && !personasError && (
-            <List>
-                {personaList.map((p) => (
-                    <ListItemButton key={p.id} selected={selectedPersona?.id === p.id} onClick={() => handleNavigation(() => handleSelectPersona(p))}>
-                        <ListItemText primary={p.name} />
-                    </ListItemButton>
-                ))}
-            </List>
-        )}
-    </Box>
-  );
-
   return (
     <ThemeProvider theme={currentTheme}>
       <CssBaseline />
@@ -1096,7 +969,7 @@ function HomePage() {
             onShowPersonas={() => handleNavigation(() => setCurrentView('personas'))}
             onShowCampaigns={() => handleNavigation(() => setCurrentView('campaigns'))}
             currentView={currentView}
-            onPersonaMenuClick={() => setPersonaDrawerOpen(!personaDrawerOpen)}
+            onPersonaMenuClick={() => { /* This is now handled by PersonasPage */ }}
         />
         {currentView === 'campaigns' && (
           <>
@@ -1203,87 +1076,7 @@ function HomePage() {
                 <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', mt: 4, px: 2 }} ><Button onClick={handleBack} disabled={activeStep === 0} variant="outlined" sx={{ borderRadius: 2, px: 3, py: 1.5 }} >Anterior</Button><Box sx={{ flexGrow: 1, display: 'flex', gap: 1, flexWrap: 'wrap', justifyContent: 'center', mx: 2 }}>{steps.map((_, index) => (<Box key={index} sx={{ width: 8, height: 8, borderRadius: '50%', backgroundColor: index === activeStep ? 'primary.main' : index < activeStep ? 'success.main' : 'grey.300', transition: 'all 0.3s ease' }} />))}</Box><Button onClick={handleNext} disabled={isGenerating || activeStep === steps.length - 1 || !canProceedToStep(activeStep + 1)} variant="contained" sx={{ borderRadius: 2, px: 3, py: 1.5 }} >Próximo</Button></Box>
               </>
             )}
-            {currentView === 'personas' && (
-              <Box sx={{ display: 'flex', width: '100%' }}>
-                  <Drawer
-                      variant={isMobile ? 'temporary' : 'persistent'}
-                      anchor="left"
-                      open={personaDrawerOpen}
-                      onClose={() => handleNavigation(() => setPersonaDrawerOpen(false))}
-                      sx={{
-                          width: 320,
-                          flexShrink: 0,
-                          '& .MuiDrawer-paper': {
-                              width: 320,
-                              boxSizing: 'border-box',
-                          },
-                      }}
-                  >
-                      <Toolbar />
-                      {personaDrawerContent}
-                  </Drawer>
-                  <Box
-                      component="main"
-                      sx={{
-                          flexGrow: 1,
-                          p: 3,
-                          transition: theme.transitions.create('margin', {
-                            easing: theme.transitions.easing.sharp,
-                            duration: theme.transitions.duration.leavingScreen,
-                          }),
-                          marginLeft: !isMobile ? `-${320}px` : 0,
-                          ...(!isMobile && personaDrawerOpen && {
-                            transition: theme.transitions.create('margin', {
-                                easing: theme.transitions.easing.easeOut,
-                                duration: theme.transitions.duration.enteringScreen,
-                            }),
-                            marginLeft: 0,
-                          }),
-                      }}
-                  >
-                      <Toolbar />
-                      <Box>
-                        {selectedPersona ? (
-                          isMobile ? (
-                            <PersonaWizard
-                              key={selectedPersona.id || 'new'}
-                              open={Boolean(selectedPersona)}
-                              onClose={() => handleNavigation(() => setSelectedPersona(null))}
-                              onSave={handleSavePersona}
-                              onReset={handleNewPersona}
-                              personaData={personaFormData}
-                              onPersonaDataChange={setPersonaFormData}
-                              onGenerate={handleGeneratePersonaWithAI}
-                              isGeneratingPersona={isGeneratingPersona}
-                              initialStep={initialWizardStep}
-                            />
-                          ) : (
-                            <Paper elevation={2} sx={{ p: 3 }}>
-                              <PersonaWizard
-                                key={selectedPersona.id || 'new'}
-                                open={Boolean(selectedPersona)}
-                                onClose={() => handleNavigation(() => setSelectedPersona(null))}
-                                onSave={handleSavePersona}
-                                onReset={handleNewPersona}
-                                personaData={personaFormData}
-                                onPersonaDataChange={setPersonaFormData}
-                                onGenerate={handleGeneratePersonaWithAI}
-                                isGeneratingPersona={isGeneratingPersona}
-                                initialStep={initialWizardStep}
-                              />
-                            </Paper>
-                          )
-                        ) : (
-                          <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center', height: '50vh' }}>
-                            <Typography variant="h6" color="text.secondary">
-                              Selecione uma persona para editar ou crie uma nova.
-                            </Typography>
-                          </Box>
-                        )}
-                      </Box>
-                  </Box>
-              </Box>
-            )}
+            {currentView === 'personas' && <PersonasPage />}
         </Box>
       </Box>
       <UnsavedChangesDialog
